setNewIndex modified

+++ TowerDefense/Tower.pde
@@ -61,7 +61,7 @@ public class Tower {
     float distance = dist(x, y, value.xCor, value.yCor);
     return (distance <= range);
   }
-  public float getRange() {
+  public int getRange() {
     return range;
   }
 
@@ -69,7 +69,7 @@ public class Tower {
     return damage;
   }
 
-  public float getDelay() {
+  public int getDelay() {
     return delay;
   }
